+++ Tests/Editor/Stubs/RecursiveWrapperService.cs
@@ -1,0 +1,13 @@
+﻿namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor.Stubs
+{
+    /// <summary>
+    /// Just a simple service which holds a reference to a recursive service.
+    /// </summary>
+    public class RecursiveWrapperService
+    {
+        public RecursiveWrapperService(RecursiveAService a)
+        {
+            
+        }
+    }
+}