+++ Tests/Editor/Stubs/ScalarService.cs
@@ -1,0 +1,10 @@
+﻿namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor.Stubs
+{
+    public class ScalarService
+    {
+        public ScalarService(int a)
+        {
+            
+        }
+    }
+}