+++ Tests/Editor/Stubs/SecondSimpleService.cs
@@ -1,0 +1,10 @@
+﻿namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor.Stubs
+{
+    /// <summary>
+    /// This is just a simple service which also extends the ISimpleInterface
+    /// </summary>
+    public class SecondSimpleService : ISimpleInterface
+    {
+        
+    }
+}