+++ Tests/Editor/Stubs/ServiceWithMultipleConstructors.cs
@@ -1,0 +1,13 @@
+﻿namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor.Stubs
+{
+    public class ServiceWithMultipleConstructors
+    {
+        public ServiceWithMultipleConstructors(SimpleService simple)
+        {
+        }
+
+        public ServiceWithMultipleConstructors(AnotherService another)
+        {
+        }
+    }
+}