+++ Tests/Editor/Stubs/SimpleEnum.cs
@@ -1,0 +1,8 @@
+﻿namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor.Stubs
+{
+    public enum SimpleEnum
+    {
+        ValueA,
+        ValueB
+    }
+}