+++ Tests/Editor/Stubs/SimpleService.cs
@@ -1,0 +1,10 @@
+﻿namespace TheRealIronDuck.Ducktion.Editor.Tests.Editor.Stubs
+{
+    /// <summary>
+    /// This class is a simple empty service without any dependencies.
+    /// </summary>
+    public class SimpleService : SimpleBaseClass, ISimpleInterface
+    {
+        
+    }
+}